+++ io_scene_habitatb/ui.py
@@ -339,7 +339,6 @@ class RevoltLightPanel(bpy.types.Panel):
                 row = box.row()
                 row.operator("lighttools.bakevertex", text="Generate Shading", icon="LIGHTPAINT")
 
-
             box = self.layout.box()
             box.label(text="Generate Shadow Texture")
             row = box.row()
@@ -350,6 +349,8 @@ class RevoltLightPanel(bpy.types.Panel):
             col.prop(context.object.revolt, "shadow_resolution")
             row = box.row()
             row.operator("lighttools.bakeshadow", icon="LAMP_SPOT", text="Generate Shadow")
+            row = box.row()
+            row.prop(context.object.revolt, "shadow_table", text="Table")
 
 # BUTTONS
 
@@ -487,16 +488,37 @@ class ButtonBakeShadow(bpy.types.Operator):
         # create a texture
         shadow_tex = bpy.data.images.new(name="Shadow", width=resolution, height=resolution)
 
-        bb = shade_obj.bound_box
-        loc = (((bb[0][0] + bb[4][0]) / 2 * shade_obj.scale[0]) + shade_obj.location[0],
-               ((bb[0][1] + bb[3][1]) / 2 * shade_obj.scale[1]) + shade_obj.location[1],
-                (bb[0][2] * shade_obj.scale[2]) + shade_obj.location[2])
+        all_objs = [ob_child for ob_child in context.scene.objects if ob_child.parent == shade_obj] + [shade_obj]
+        print([ob.matrix_local for ob in all_objs])
+
+        # get the bounds taking in account all child objects (wheels, etc.)
+        far_left = min([min([(ob.matrix_world[0][3] + ob.bound_box[i][0] * shade_obj.scale[0])  for i in range(0, 8)]) for ob in all_objs])
+        far_right = max([max([(ob.matrix_world[0][3] + ob.bound_box[i][0] * shade_obj.scale[0])  for i in range(0, 8)]) for ob in all_objs])
+        far_front = max([max([(ob.matrix_world[1][3] + ob.bound_box[i][1] * shade_obj.scale[1])  for i in range(0, 8)]) for ob in all_objs])
+        far_back = min([min([(ob.matrix_world[1][3] + ob.bound_box[i][1] * shade_obj.scale[1])  for i in range(0, 8)]) for ob in all_objs])
+        far_top = max([max([(ob.matrix_world[2][3] + ob.bound_box[i][2] * shade_obj.scale[2])  for i in range(0, 8)]) for ob in all_objs])
+        far_bottom = min([min([(ob.matrix_world[2][3] + ob.bound_box[i][2] * shade_obj.scale[2])  for i in range(0, 8)]) for ob in all_objs])
+
+        # get the dimensions to set the scale
+        dim_x = abs(far_left - far_right)
+        dim_y = abs(far_front - far_back)
+
+        # location for the shadow plane
+        loc = ((far_right + far_left)/2,
+               (far_front + far_back)/2,
+                far_bottom)
 
         # create the shadow plane and map it
         bpy.ops.mesh.primitive_plane_add(location=loc, enter_editmode=True)
         bpy.ops.uv.unwrap()
         bpy.ops.object.mode_set(mode='OBJECT')
         shadow_plane = context.object
+
+        # scale the shadow plane
+        scale = max(dim_x, dim_y)
+        shadow_plane.scale[0] = scale/1.5
+        shadow_plane.scale[1] = scale/1.5
+        print(shadow_plane.scale)
 
         for uv_face in context.object.data.uv_textures.active.data:
             uv_face.image = shadow_tex
@@ -513,6 +535,19 @@ class ButtonBakeShadow(bpy.types.Operator):
         # select the other object again
         shade_obj.select = True
         scene.objects.active = shade_obj
+
+        # space between the car body center and the edge of the shadow plane
+        sphor = (shadow_plane.location[0] - (shadow_plane.dimensions[0]/2))
+        spver = ((shadow_plane.dimensions[1]/2) - shadow_plane.location[1])
+
+        # generate shadowtable
+        sleft = (sphor - shade_obj.location.x) * 100
+        sright = (shade_obj.location[0] - sphor) * 100
+        sfront = (spver - shade_obj.location[1]) * 100
+        sback = (shade_obj.location[1] - spver) * 100
+        sheight = (far_bottom - shade_obj.location[2]) * 100
+        shtable = ";)SHADOWTABLE {:4f} {:4f} {:4f} {:4f} {:4f}".format(sleft, sright, sfront, sback, sheight)
+        shade_obj.revolt.shadow_table = shtable
 
         return{'FINISHED'}