+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog HabitatB
 
 ## 2017-05-08
-
 + Car textures will now be loaded even if / is used instead of \ in parameters.txt (thanks to Gotolei for reporting)
 + Textures are now kept even after reopening a .blend file with an imported prm/w
 + Export now takes object scale, rotation and translation into account properly. This should also fix issues with exported ncp files. (Thanks to R6 for reporting)
@@ -10,7 +9,6 @@
 + Many small bug-fixes that you can read about on GitHub.
 
 ## 2017-07-26
-
 + Experimental POS support (only import, no export)
 + Fixed the PRM export issue where only 50% of the faces were exported
 + Fixed the selection buttons for faces in the tools panel.
@@ -23,7 +21,13 @@
 ## 2017-07-28
 + Added Lights and Shadow section in the Tool panel
 + Added button for automatically shading objects
+
 ## 2017-07-29
 + Added more options for shading automatically
 + Layout updates and fixes for the automatic shading
 + Added a function for baking inverted shadows automatically
+
+## 2017-07-30
++ Added options to the shadow generator
++ Shadowtable generator Added
++ UI cleanup